requirements.txt
alembic==1.17.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.0.1
mysql==0.0.3
mysqlclient==2.2.7
packaging==25.0
py==1.11.0
PyMySQL==1.1.2
python-dotenv==1.1.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
